Test empty value for `(min)` and `(max)`
